+++ src/internal/MicrogrammarSpecParser.ts
@@ -1,13 +1,14 @@
-import { Concat } from "./Concat";
-import { Config, DefaultConfig } from "./Config";
-import { InputState } from "./InputState";
-import { Term } from "./Matchers";
-import { MatchPrefixResult } from "./MatchPrefixResult";
-import { Microgrammar } from "./Microgrammar";
-import { isPatternMatch } from "./PatternMatch";
-import { Literal, Regex } from "./Primitives";
-import { Rep } from "./Rep";
-import { Break } from "./snobol/Snobol";
+import { Concat } from "../Concat";
+import { Config, DefaultConfig } from "../Config";
+import { Term } from "../Matchers";
+import { MatchPrefixResult } from "../MatchPrefixResult";
+import { Microgrammar } from "../Microgrammar";
+import { isPatternMatch } from "../PatternMatch";
+import { Literal, Regex } from "../Primitives";
+import { Rep } from "../Rep";
+import { Break } from "../snobol/Snobol";
+
+import { inputStateFromString } from "./InputStateFactory";
 
 /**
  * Parses microgrammars expressed as strings.
@@ -34,7 +35,7 @@ export class MicrogrammarSpecParser {
             trailing: new Break("this will not appear"), // replace with matchEverything after merge
         } as Term);
 
-        const mpr: MatchPrefixResult = specGrammar.matchPrefix(InputState.fromString(spec), {});
+        const mpr: MatchPrefixResult = specGrammar.matchPrefix(inputStateFromString(spec), {});
         if (!isPatternMatch(mpr)) {
             throw new Error(`Unable to parse microgrammar: ${spec}`);
         }