+++ test/MicrogrammarTest.ts
@@ -3,7 +3,7 @@ import * as assert from "power-assert";
 import { MatchingLogic, Term } from "../src/Matchers";
 import { MatchingMachine, Microgrammar } from "../src/Microgrammar";
 import { Opt } from "../src/Ops";
-import { PatternMatch } from "../src/PatternMatch";
+import {isPatternMatch, PatternMatch} from "../src/PatternMatch";
 import { Rep1Sep, RepSep } from "../src/Rep";
 import { RealWorldPom } from "./Fixtures";
 import {
@@ -12,8 +12,10 @@ import {
 } from "./MavenGrammars";
 
 import { fail } from "power-assert";
+import { JavaParenthesizedExpression } from "../src/java/JavaBody";
+import { JAVA_IDENTIFIER } from "./java/JavaBlockMicrogrammarTest";
 
-describe("MicrogrammarTest", () => {
+describe("Microgrammar", () => {
 
     it("literal", () => {
         const content = "foo ";
@@ -54,6 +56,74 @@ describe("MicrogrammarTest", () => {
         } catch (e) {
             assert(e.toString().lastIndexOf("content") !== -1);
         }
+    });
+
+    it("parse all content: File matches", () => {
+        const content = "public void thing(int i);";
+        const mg = Microgrammar.fromDefinitions<{ name: string }>({
+            _p: "public",
+            type: JAVA_IDENTIFIER,
+            name: JAVA_IDENTIFIER,
+            params: JavaParenthesizedExpression,
+            _semi: ";",
+        });
+        const result = mg.exactMatch(content);
+        if (isPatternMatch(result)) {
+            assert(result);
+            assert(result.$matched === content);
+            assert(result.name === "thing");
+        } else {
+            fail();
+        }
+    });
+
+    it("parse all content: pattern match recognized in output", () => {
+        const content = "public void";
+        const mg = Microgrammar.fromDefinitions<any>({
+            _p: "public",
+            type: JAVA_IDENTIFIER,
+        });
+        const result = mg.exactMatch(content);
+        assert(isPatternMatch(result));
+    });
+
+    it("parse all content: dismatch report recognized in output", () => {
+        const content = "not-matchy void";
+        const mg = Microgrammar.fromDefinitions<{type: string }>({
+            _p: "public",
+            type: JAVA_IDENTIFIER,
+        });
+        const result = mg.exactMatch(content);
+        assert(!isPatternMatch(result));
+        if (!isPatternMatch(result)) {
+            assert(result.description !== undefined);
+        }
+    });
+
+    it("parse all content: Fail due to irrelevant content after match", () => {
+        const content = "public void thing(int i); // and this is irrelevant crap";
+        const mg = Microgrammar.fromDefinitions<any>({
+            _p: "public",
+            type: JAVA_IDENTIFIER,
+            name: JAVA_IDENTIFIER,
+            params: JavaParenthesizedExpression,
+            _semi: ";",
+        });
+        const result = mg.exactMatch(content);
+        assert(!isPatternMatch(result));
+    });
+
+    it("parse all content: Fail due to irrelevant content before match", () => {
+        const content = "// and this is irrelevant crap\npublic void thing(int i);";
+        const mg = Microgrammar.fromDefinitions<any>({
+            _p: "public",
+            type: JAVA_IDENTIFIER,
+            name: JAVA_IDENTIFIER,
+            params: JavaParenthesizedExpression,
+            _semi: ";",
+        });
+        const result = mg.exactMatch(content);
+        assert(!isPatternMatch(result));
     });
 
     it("can JSON stringify", () => {