+++ test/WhenTest.ts
@@ -1,14 +1,14 @@
 import { expect } from "chai";
-import { InputState } from "../src/InputState";
+import { inputStateFromString } from "../src/internal/InputStateFactory";
 import { when } from "../src/Ops";
 import { isPatternMatch } from "../src/PatternMatch";
 import { Literal } from "../src/Primitives";
 
-describe("Whentest", () => {
+describe("When", () => {
 
     it("when true should match", () => {
         const primitive = new Literal("foo");
-        const is = InputState.fromString("foo bar");
+        const is = inputStateFromString("foo bar");
         const matcher = when(primitive, pm => true);
         if (!matcher) {
             throw new Error("Error: matcher returned by when is undefined");
@@ -22,7 +22,7 @@ describe("Whentest", () => {
 
     it("when false should not match", () => {
         const primitive = new Literal("foo");
-        const is = InputState.fromString("foo bar");
+        const is = inputStateFromString("foo bar");
         const matcher = when(primitive, pm => false);
         const m = matcher.matchPrefix(is);
         expect(isPatternMatch(m)).to.equal(false);
@@ -30,7 +30,7 @@ describe("Whentest", () => {
 
     it("ability to veto content", () => {
         const primitive = new Literal("foo");
-        const is = InputState.fromString("foo bar");
+        const is = inputStateFromString("foo bar");
         const hatesFoo = when(primitive, pm => pm.$matched.indexOf("foo") === -1);
         const m = hatesFoo.matchPrefix(is);
         expect(isPatternMatch(m)).to.equal(false);
@@ -38,7 +38,7 @@ describe("Whentest", () => {
 
     it("ability to veto content: not vetoed", () => {
         const primitive = new RegExp(/^[a-z]+/);
-        const is = InputState.fromString("bar and this is a load of other stuff");
+        const is = inputStateFromString("bar and this is a load of other stuff");
         const hatesFoo = when(primitive, pm => pm.$matched.indexOf("foo") === -1);
         const m = hatesFoo.matchPrefix(is);
         expect(isPatternMatch(m)).to.equal(true);
@@ -46,7 +46,7 @@ describe("Whentest", () => {
 
     it("ability to require content: match", () => {
         const primitive = new Literal("foo");
-        const is = InputState.fromString("foo bar");
+        const is = inputStateFromString("foo bar");
         const requiresFoo = when(primitive, pm => pm.$matched.indexOf("foo") !== -1);
         const m = requiresFoo.matchPrefix(is);
         expect(isPatternMatch(m)).to.equal(true);
@@ -54,7 +54,7 @@ describe("Whentest", () => {
 
     it("ability to require content: no match", () => {
         const primitive = new Literal("foo");
-        const is = InputState.fromString("bar");
+        const is = inputStateFromString("bar");
         const requiresFoo = when(primitive, pm => pm.$matched.indexOf("foo") !== -1);
         const m = requiresFoo.matchPrefix(is);
         expect(isPatternMatch(m)).to.equal(false);