+++ test/internal/InputStateTest.ts
@@ -1,37 +1,38 @@
 import { expect } from "chai";
-import { InputState, InputStateManager } from "../src/InputState";
-import { InputStream } from "../src/InputStream";
-import { DEPENDENCY_GRAMMAR } from "./MavenGrammars";
+
+import { InputStream } from "../../src/spi/InputStream";
+import { DEPENDENCY_GRAMMAR } from "../MavenGrammars";
 
 import * as assert from "power-assert";
+import { inputStateFromStream, inputStateFromString } from "../../src/internal/InputStateFactory";
 
-describe("InputStateTest", () => {
+describe("InputState", () => {
 
     it("cannot consume", () => {
-        const is = InputState.fromString("foo bar");
+        const is = inputStateFromString("foo bar");
         assert.throws(
             () => is.consume("xxxx"),
         );
     });
 
     it("remainder is all", () => {
-        const is = InputState.fromString("foo bar");
+        const is = inputStateFromString("foo bar");
         expect(is.peek(10)).equals("foo bar");
     });
 
     it("remainder is correct after advance", () => {
-        const is = InputState.fromString("foo bar").advance();
+        const is = inputStateFromString("foo bar").advance();
         // expect(is.content).equals("foo bar");
         expect(is.peek(1000)).equals("oo bar");
     });
 
     it("remainder is correct after attempt advance past end", () => {
-        const is = InputState.fromString("f").advance();
+        const is = inputStateFromString("f").advance();
         expect(is.peek(10)).equals("");
     });
 
     it("peek is correct after read", () => {
-        const is = InputState.fromString("0123456789");
+        const is = inputStateFromString("0123456789");
         assert(is.peek(1) === "0");
         const at3 = is.consume("012");
         assert(is.peek(1) === "0");
@@ -71,7 +72,7 @@ describe("InputStateTest", () => {
         withBuffer(300));
 
     function withBuffer(bufSize: number, extraContent: string = "") {
-        let is = InputState.fromString("0123456789" + extraContent);
+        let is = inputStateFromString("0123456789" + extraContent);
         assert(is.peek(7) === "0123456");
         is = is.consume("01");
         assert(is.peek(3) === "234");
@@ -82,9 +83,7 @@ describe("InputStateTest", () => {
     it("read ahead does not dirty parent", () => {
         const stream = new ReleasingStringInputStream("the quick brown fox jumps over the lazy dog");
         expect(stream.offset).to.equal(0);
-
-        const state0 = new InputState(new InputStateManager(stream));
-
+        const state0 = inputStateFromStream(stream);
         const state1 = state0.advance();
         expect(state1.offset).to.equal(1);
         const state2 = state1.advance();